Automate default config delivering.

+++ deliver_resources.sh
@@ -2,6 +2,7 @@
 
 buildFolder="Build-cmake"
 fileName="spellbook.json"
+fileNameDefault="spellbook_default.json"
 databaseFile="resources/$fileName"
 
 curDir=$PWD
@@ -13,10 +14,20 @@ if [ ! -z "$1" ]; then
 	echo "Now in '$1'"
 fi
 
+initial=false
+if [ "$2" == "-i" ]; then
+	initial=true
+fi
 destPath="$buildFolder/Spellbook"
 destFilePath="$destPath/$fileName"
-rm "$destFilePath"
-echo "Copy '$databaseFile' to the build folder $destPath/"
-cp "$databaseFile" "$destFilePath"
+[ -f "$destFilePath" ] && rm "$destFilePath"
+
+if ! $initial; then
+	echo "Copy '$databaseFile' to the build folder $destPath/"
+	cp "$databaseFile" "$destFilePath"
+fi
+destFilePathDefault="$destPath/$fileNameDefault"
+echo "Create default config in '$destFilePathDefault'"
+cp "$databaseFile" "$destFilePathDefault"
 
 exit 0
